Refactor ViewModelTest
테스트시 데이터 셋업과 정리를 ViewModel에서 수행할 수 있도록 수정했다.
